refactor(github): update event name & function id
