+++ apps/github/src/app/api/webhook/sync-organisation/service.test.ts
@@ -37,7 +37,7 @@ describe('handleElbaOrganisationActivated', () => {
     });
     expect(send).toBeCalledTimes(1);
     expect(send).toBeCalledWith({
-      name: 'third-party-apps/sync',
+      name: 'third-party-apps/page_sync.requested',
       data: {
         organisationId: organisation.id,
         installationId: organisation.installationId,
