+++ apps/github/src/inngest/functions/index.ts
@@ -1,6 +1,6 @@
 import { scheduleAppsSyncs } from './third-party-apps/schedule-apps-syncs';
-import { syncApps } from './third-party-apps/sync-apps';
+import { syncApps } from './third-party-apps/sync-apps-page';
 import { scheduleUsersSyncs } from './users/schedule-users-syncs';
-import { syncUsers } from './users/sync-users';
+import { syncUsers } from './users/sync-users-page';
 
 export const inngestFunctions = [syncUsers, scheduleUsersSyncs, syncApps, scheduleAppsSyncs];
